Separate tests by the usage of iptables
