+++ .github/workflows/go.yml
@@ -1,4 +1,4 @@
-name: Go CI
+name: Yetis Test
 
 on:
   push:
@@ -6,7 +6,23 @@ on:
       - '*'
 
 jobs:
-  test:
+  tests_without_iptables:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v3
+
+      # Set up Docker
+      - name: Set up Go
+        uses: actions/setup-go@v4
+        with:
+          go-version: '1.23'
+
+      - name: Run Go tests
+        run: make test_skip_iptables
+
+  tests_with_iptables:
     runs-on: ubuntu-latest
 
     steps:
@@ -17,5 +33,5 @@ jobs:
       - name: Set up Docker Buildx
         uses: docker/setup-buildx-action@v2
 
-      - name: Run Go tests
-        run: make test
+      - name: Run tests requiring iptables
+        run: make test_only_iptables
