+++ Makefile
@@ -3,7 +3,15 @@ compile:
 	GOOS=linux GOARCH=amd64 go build -o yetis main.go && chmod +x yetis && mv yetis build/
 
 test:
-	docker run --privileged --rm -v $$PWD:/usr/src/myapp -w /usr/src/myapp -v /var/run/docker.sock:/var/run/docker.sock glossd/goyetis:0.3-amd go test -v ./...
+	make test_skip_iptables
+	make test_only_iptables
+
+test_skip_iptables:
+	go test ./...
+
+test_only_iptables:
+	docker run --privileged --rm -v $$PWD:/usr/src/myapp -w /usr/src/myapp -v /var/run/docker.sock:/var/run/docker.sock glossd/goyetis:0.3-amd\
+ sh -c "export TEST_IPTABLES=true; go test -v ./itests/iptables_test.go; go test -v ./proxy/..."
 
 docker-sh:
 	docker run -it --privileged --rm -v $$PWD:/usr/src/myapp -w /usr/src/myapp glossd/goyetis:0.3.1 bash
